Update index.html
I made couple of changes to both the html and the CSS sides of the project
I created a wrapper class to wrap around all items at the center of the page.
The nav-bar was also updated as well as the member class. Grid was used to position the the member photos at the center of the page.